+++ login_with_amazon/example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 import 'package:login_with_amazon/login_with_amazon.dart';
 
@@ -13,13 +14,13 @@ class MyApp extends StatefulWidget {
 const _scopes = {
   'profile': null,
   'alexa:all': {
-    'productID': 'SomeProductId',
+    'productID': 'YOURPRODUCTID',
     'productInstanceAttributes': {
       'deviceSerialNumber': 'serialNumberHere',
     },
   },
   'dash:replenish': {
-    'device_model': 'SomeDeviceModel',
+    'device_model': 'YOURDEVICEMODEL',
     'serial': 'serialNumberHere',
     'is_test_device': true
   },
@@ -33,9 +34,14 @@ class _MyAppState extends State<MyApp> {
   }
 
   List<Widget> _authRespose = [];
-  String _accessToken = '';
-  String _authCode = '';
   String _logResult = '';
+  String _result = '';
+  String _codeChallengeMethod = 'plain';
+  final _alexaId = TextEditingController(text: 'YOURPRODUCTID');
+  final _dashModel = TextEditingController(text: 'YOURDEVICEMODEL');
+  final _serialNumber = TextEditingController(text: 'someSerialNeeded');
+  final _codeChallenge = TextEditingController(
+      text: 'code_generated_by_the_dash_device_for_LWA_verification');
 
   @override
   Widget build(BuildContext context) {
@@ -45,12 +51,54 @@ class _MyAppState extends State<MyApp> {
           title: const Text('Plugin example app'),
         ),
         body: Center(
-          child: Column(
+          child: ListView(
             children: <Widget>[
+              TextFormField(
+                controller: _alexaId,
+                decoration: InputDecoration(
+                  labelText: 'Alexa Product Id',
+                ),
+              ),
+              TextFormField(
+                controller: _dashModel,
+                decoration: InputDecoration(
+                  labelText: 'Dash Model',
+                ),
+              ),
+              TextFormField(
+                controller: _serialNumber,
+                decoration: InputDecoration(
+                  labelText: 'Serial Number',
+                ),
+              ),
+              TextFormField(
+                controller: _codeChallenge,
+                decoration: InputDecoration(
+                  labelText: 'Code Challenge',
+                ),
+              ),
+              DropdownButtonFormField(
+                  onChanged: (val) {
+                    setState(() {
+                      _codeChallengeMethod = val;
+                    });
+                  },
+                  value: _codeChallengeMethod,
+                  items: ['plain', 'S256']
+                      .map((type) => DropdownMenuItem(
+                            value: type,
+                            child: Text(type),
+                          ))
+                      .toList()),
               RaisedButton(
                 onPressed: () async {
                   try {
-                    final result = await LoginWithAmazon.login(_scopes);
+                    final result = await LoginWithAmazon.login({
+                      'profile': null,
+                    });
+                    _result = result.entries
+                        .map((entry) => "${entry.key}: ${entry.value}")
+                        .reduce((str, str2) => "$str, \n $str2");
                     setState(() {
                       _logResult = result.containsKey("accessToken")
                           ? "Logged In"
@@ -61,7 +109,7 @@ class _MyAppState extends State<MyApp> {
                     });
                   } catch (e) {
                     setState(() {
-                      _logResult = "Cancelled";
+                      _logResult = "failed";
                     });
                   }
                 },
@@ -70,33 +118,60 @@ class _MyAppState extends State<MyApp> {
               RaisedButton(
                 onPressed: () async {
                   final result = await LoginWithAmazon.getAuthCode(
-                    "code_generated_by_the_dash_device_for_LWA_verification",
+                    _codeChallenge.text,
                     "plain",
                     {
                       'profile': null,
                       'dash:replenish': {
-                        'device_model': 'SomeDeviceModel',
-                        'serial': 'serialNumberHere',
+                        'device_model': _dashModel.text,
+                        'serial': _serialNumber.text,
                         'is_test_device': true
                       },
                     },
                   );
+                  _result = result.entries
+                      .map((entry) => "${entry.key}: ${entry.value}")
+                      .reduce((str, str2) => "$str, \n $str2");
                   setState(() {
-                    _authCode = result.containsKey("authorizationCode")
-                        ? result["authorizationCode"]
-                        : "None";
                     _authRespose = result.entries.map((entry) {
                       return Text("${entry.key}: ${entry.value}");
                     }).toList();
                   });
                 },
-                child: Text("Get Auth Code"),
+                child: Text("Get Dash Auth Code"),
+              ),
+              RaisedButton(
+                onPressed: () async {
+                  final result = await LoginWithAmazon.getAuthCode(
+                    "code_generated_by_the_dash_device_for_LWA_verification",
+                    "plain",
+                    {
+                      'profile': null,
+                      'alexa:all': {
+                        'productID': _alexaId.text,
+                        'productInstanceAttributes': {
+                          'deviceSerialNumber': _serialNumber.text,
+                        },
+                      }
+                    },
+                  );
+                  _result = result.entries
+                      .map((entry) => "${entry.key}: ${entry.value}")
+                      .reduce((str, str2) => "$str, \n $str2");
+                  setState(() {
+                    _authRespose = result.entries.map((entry) {
+                      return Text("${entry.key}: ${entry.value}");
+                    }).toList();
+                  });
+                },
+                child: Text("Get Alexa Auth Code"),
               ),
               RaisedButton(
                 onPressed: () async {
                   final result = await LoginWithAmazon.getAccessToken(_scopes);
                   setState(() {
-                    _accessToken = result;
+                    _authRespose = [Text('AccessToken: $result')];
+                    _result = result;
                   });
                 },
                 child: Text("Get Access Token"),
@@ -106,14 +181,19 @@ class _MyAppState extends State<MyApp> {
                   await LoginWithAmazon.logout();
                   setState(() {
                     _logResult = 'Logged Out';
+                    _result = '';
                   });
                 },
                 child: Text("Logout"),
               ),
               Text('LogResult: $_logResult'),
-              Text('AccessToken: $_accessToken'),
-              Text('AuthCode: $_authCode'),
-              ..._authRespose
+              ..._authRespose,
+              RaisedButton(
+                onPressed: () async {
+                  Clipboard.setData(ClipboardData(text: _result));
+                },
+                child: Text("Copy To Clipboard"),
+              )
             ],
           ),
         ),